Use touch mode for probing to workaround problems with clones
It seems like we are normally not fast enough to retract the pin,
and that causes BLTouch probes to misbehave. Using touch mode bypasses
that.

However it looks like the touch mode has much, much worse repeatability,
an order of magnitude worse, so this is only a band aid sollution.

Signed-off-by: Fred Sundvik <fsundvik@gmail.com>