Don't continously send the last BLTouch command
There was a missing command to reset the PWM pulse after sending. This
meant that the last command was continously sent. This also meant that
the following command could randomly fail.

This commit also removes the previous band-aid solution of using touch
mode, since it does no longer seem to be needed

Signed-off-by: Fred Sundvik <fsundvik@gmail.com>