Home with multiple probes
Klipper does not really support this natively, if it would, then we
would most likely probe for say 10 times and use the median result.

As a workaround it uses endstop_phase, with just the accuracy set. So if
one of the probes differs too much from the others it will fail.